Some controls tab performance improvements
