+++ src/hooks/useFlowChartState.ts
@@ -2,7 +2,6 @@ import { ElementsData } from "@src/feature/flow_chart_panel/types/CustomNodeProp
 import { atom, useAtom } from "jotai";
 import { atomWithImmer } from "jotai-immer";
 import localforage from "localforage";
-import { Layout } from "react-grid-layout";
 import { ReactFlowJsonObject } from "reactflow";
 
 export interface CtrlManifestParam {
@@ -45,39 +44,14 @@ export interface RfSpatialInfoType {
   zoom: number;
 }
 
-const initialManifests: CtlManifestType[] = [
-  {
-    type: "input",
-    name: "Slider",
-    id: "INPUT_PLACEHOLDER",
-    hidden: false,
-    minHeight: 1,
-    minWidth: 2,
-    layout: {
-      x: 0,
-      y: 0,
-      h: 2,
-      w: 2,
-      minH: 1,
-      minW: 2,
-      i: "INPUT_PLACEHOLDER",
-    },
-  },
-];
 const failedNodeAtom = atomWithImmer<string>("");
 const runningNodeAtom = atomWithImmer<string>("");
 const showLogsAtom = atomWithImmer<boolean>(false);
 const rfInstanceAtom = atomWithImmer<
   ReactFlowJsonObject<ElementsData> | undefined
 >(undefined);
-const manifestAtom = atomWithImmer<CtlManifestType[]>(initialManifests);
 const editModeAtom = atomWithImmer<boolean>(false);
 const expandModeAtom = atomWithImmer<boolean>(false);
-const gridLayoutAtom = atomWithImmer<Layout[]>(
-  initialManifests.map((ctrl) => ({
-    ...ctrl.layout,
-  }))
-);
 const apiKeyAtom = atomWithImmer<string>("");
 const isSidebarOpenAtom = atom<boolean>(false);
 const nodeParamChangedAtom = atom<boolean | undefined>(undefined);
@@ -85,10 +59,8 @@ localforage.config({ name: "react-flow", storeName: "flows" });
 
 export function useFlowChartState() {
   const [rfInstance, setRfInstance] = useAtom(rfInstanceAtom);
-  const [ctrlsManifest, setCtrlsManifest] = useAtom(manifestAtom);
   const [isEditMode, setIsEditMode] = useAtom(editModeAtom);
   const [isExpandMode, setIsExpandMode] = useAtom(expandModeAtom);
-  const [gridLayout, setGridLayout] = useAtom(gridLayoutAtom);
   const [showLogs, setShowLogs] = useAtom(showLogsAtom);
   const [runningNode, setRunningNode] = useAtom(runningNodeAtom);
   const [failedNode, setFailedNode] = useAtom(failedNodeAtom);
@@ -99,14 +71,10 @@ export function useFlowChartState() {
   return {
     rfInstance,
     setRfInstance,
-    ctrlsManifest,
-    setCtrlsManifest,
     isEditMode,
     setIsEditMode,
     isExpandMode,
     setIsExpandMode,
-    gridLayout,
-    setGridLayout,
     showLogs,
     setShowLogs,
     runningNode,
